requirements.txt
tensorflow==2.16.2
opencv-python==4.7.0.68
matplotlib==3.9.4
numpy==1.26.4
h5py==3.14.0
Pillow==11.3.0
mediapipe==0.10.21